webapp/requirements.txt
Django==1.11
django-elasticsearch-dsl==0.4.4
django-elasticsearch-dsl-drf==0.6.3
django-nine==0.1.13
djangorestframework==3.7.7
elasticsearch==5.5.2
elasticsearch-dsl==5.4.0
psycopg2==2.7.4
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
urllib3==1.22
